Move CreatedSwap to cnd
The struct is purely a cnd construct, nothing in `comit` relies on it
so it can be safely moved over.
